+++ plugins/module_utils/validate/base.py
@@ -40,7 +40,7 @@ def load_validator(
         return None, result
 
     cref = dict(zip(["corg", "cname", "plugin"], engine.split(".")))
-    validatorlib = "ansible_collections.{corg}.{cname}.plugins.validate.{plugin}".format(
+    validatorlib = "ansible_collections.{corg}.{cname}.plugins.sub_plugins.validate.{plugin}".format(
         **cref
     )
