+++ tests/unit/plugins/sub_plugins/cli_parsers/fixtures/nxos_show_version.ttp
